Rename module: 'tables' -> 'tabular'
