+++ tests/tabular_tests/base_tests/strategies.py
