+++ tests/tabular_tests/records_tests/test_lists_factory.py
@@ -5,7 +5,7 @@ from hypothesis import given
 from sqlalchemy.schema import Table
 
 from hypothesis_sqlalchemy.hints import Strategy
-from hypothesis_sqlalchemy.tables.records import lists_factory
+from hypothesis_sqlalchemy.tabular.records import lists_factory
 from tests.utils import (DataObject,
                          table_record_is_valid)
 from . import strategies
